Move away DerivePublicKeys from NameLedger
